avapager: Refactor pager to prevent hanging pointers and deadlocks

+++ include/avadb/avainterface.h
@@ -11,10 +11,10 @@ typedef struct AvaOSInterface {
     uint16_t version; /* Currently set to 1 */
     int (*open)(struct AvaOSInterface*, const char*, uint16_t, AvaFile**);
     int (*close)(struct AvaOSInterface*, AvaFile*);
-    int (*read)(struct AvaOSInterface*, AvaFile*, void*, int, uint64_t);
-    int (*write)(struct AvaOSInterface*, AvaFile*, void*, int, uint64_t);
-    int (*truncate)(struct AvaOSInterface*, AvaFile*, uint64_t);
-    int (*getsize)(struct AvaOSInterface*, AvaFile*);
+    size_t (*read)(struct AvaOSInterface*, AvaFile*, void*, size_t, uint64_t);
+    size_t (*write)(struct AvaOSInterface*, AvaFile*, void*, size_t, uint64_t);
+    size_t (*truncate)(struct AvaOSInterface*, AvaFile*, size_t);
+    size_t (*getsize)(struct AvaOSInterface*, AvaFile*);
     int (*sync)(struct AvaOSInterface*, AvaFile*);
     int (*lock)(struct AvaOSInterface*, AvaFile*);
     int (*unlock)(struct AvaOSInterface*, AvaFile*);
